#3564 Require an admin to be attributed to every IP ban
created_by is now NOT NULL - every ban must be traceable to the admin who
made it, including CLI bans. bannedipaddress:ban now takes a required
adminId argument, validated against the admin role, instead of silently
recording no one.

+++ app/Http/Controllers/AdminTools/AdminToolsBannedIpAddressController.php
@@ -32,7 +32,7 @@ class AdminToolsBannedIpAddressController extends Controller
             $request->string('ip_address')->toString(),
             $request->string('reason')->toString() === '' ? null : $request->string('reason')->toString(),
             $request->filled('expires_at') ? Carbon::parse($request->string('expires_at')->toString()) : null,
-            Auth::id(),
+            (int)Auth::id(),
         );
 
         Session::flash('status', __('controller.admintools.flash.banned_ip_address_added'));
